fix: reload language server client on restart
Ensure the client watcher and output channel are disposed during a
restart before the new client is creates, and also show a helper warning
when certain variables are changed that require a restart.

Also catch errors starting the client, so that our restart configuration
change handler is still registered in case the user fixes their
serverPath or something like that. Similarly, always register the
restart command in case user wants to start using language server after
not using it before.

Closes #419